# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: Система управления банковскими картами
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  "/api/cards":
    get:
      tags:
        - Карты
      summary: Получить все карты (только ADMIN)
      description: Возвращает список всех карт в системе с пагинацией
      operationId: getAll_1
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: id
        - name: direction
          in: query
          required: false
          schema:
            type: string
            default: asc
      responses:
        '200':
          description: Список всех карт
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelCardDto"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelCardDto"
      security:
        - bearerAuth: []
    put:
      tags:
        - Карты
      summary: Создать новую карту (только ADMIN)
      description: Создает новую банковскую карту для пользователя
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateCardRequest"
        required: true
      responses:
        '200':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CreateCardResponse"
        '400':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CreateCardResponse"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CreateCardResponse"
      security:
        - bearerAuth: []
...
  "/api/cards/{id}/block":
    post:
      tags:
        - Карты
      summary: Заблокировать карту
      description: Блокирует указанную карту
      operationId: block
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Карта заблокирована
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Карта истекла
          content:
            application/json:
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                type: object
      security:
        - bearerAuth: []
  "/api/cards/{id}/activate":
    post:
      tags:
        - Карты
      summary: Активировать карту (только ADMIN)
      description: Активирует указанную карту
      operationId: activate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Карта активирована
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Карта истекла
          content:
            application/json:
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                type: object
      security:
        - bearerAuth: []
...
  "/api/transactions/card/{cardId}":
    get:
      tags:
        - Транзакции
      summary: Получить транзакции по карте
      description: Возвращает список транзакций для указанной карты
      operationId: getByCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: date
        - name: direction
          in: query
          required: false
          schema:
            type: string
            default: desc
      responses:
        '200':
          description: Список транзакций карты
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelTransactionDto"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelTransactionDto"
      security:
        - bearerAuth: []
  "/api/cards/{id}":
    get:
      tags:
        - Карты
      summary: Получить карту по ID (только ADMIN)
      description: Возвращает информацию о конкретной карте
      operationId: get_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о карте
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CardDto"
        '400':
          description: Карта не найдена
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CardDto"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CardDto"
      security:
        - bearerAuth: []
    delete:
      tags:
        - Карты
      summary: Удалить карту (только ADMIN)
      description: Удаляет указанную карту из системы
      operationId: delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Карта удалена
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Карта уже удалена
          content:
            application/json:
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                type: object
      security:
        - bearerAuth: []
  "/api/cards/own":
    get:
      tags:
        - Карты
      summary: Получить собственные карты
      description: Возвращает список карт текущего пользователя с пагинацией
      operationId: own_1
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: id
        - name: direction
          in: query
          required: false
          schema:
            type: string
            default: asc
      responses:
        '200':
          description: Список карт пользователя
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelCardDto"
      security:
        - bearerAuth: []
components:
  schemas:
...
    CardDto:
      type: object
      description: DTO для представления банковской карты
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
        number:
          type: string
          description: Номер карты
        holder:
          type: string
          description: Держатель карты
        expiryDate:
          type: string
          format: date
          description: Дата окончания действия
        status:
          type: string
          description: Статус карты
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
        balance:
          type: number
          description: Баланс карты
        deleted:
          type: boolean
          description: Флаг удаления карты
    CreateCardResponse:
      type: object
      description: Ответ на создание карты
      properties:
        number:
          type: string
          description: Номер карты
          example: 1234567812345678
        card:
          "$ref": "#/components/schemas/CardDto"
          description: Информация о созданной карте
    CreateCardRequest:
      type: object
      description: Запрос на создание карты
      properties:
        holder:
          type: string
          description: Держатель карты
          example: IVAN IVANOV
          maxLength: 100
          minLength: 0
        user:
          type: integer
          format: int64
          description: ID владельца карты
          example: 1
        expiry_date:
          type: string
          format: date
          description: Дата окончания действия карты
          example: '2025-12-31'
      required:
        - expiry_date
        - holder
        - user
    TransactionDto:
      type: object
      description: DTO для представления транзакции
      properties:
        id:
          type: integer
          format: int64
          description: ID транзакции
        from:
          type: integer
          format: int64
          description: ID карты отправителя
        to:
          type: integer
          format: int64
          description: ID карты получателя
        amount:
          type: number
          description: Сумма транзакции
        comment:
          type: string
          description: Комментарий к транзакции
        date:
          type: string
          format: date-time
          description: Дата и время транзакции
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
...
    PagedModelCardDto:
      type: object
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/CardDto"
        page:
          "$ref": "#/components/schemas/PageMetadata"
    PagedModelTransactionDto:
      type: object
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/TransactionDto"
        page:
          "$ref": "#/components/schemas/PageMetadata"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT